requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
scipy>=1.12.0
httpx>=0.27.0
web3>=7.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
